+++ pkgs/remapSourcePathPrefix.nix
@@ -1,7 +1,0 @@
-{ makeSetupHook
-}:
-
-makeSetupHook
-{
-  name = "remapSourcePathPrefixHook";
-} ./remapSourcePathPrefixHook.sh